Revert integration tests for video
