+++ compositor_integration_tests/src/lib.rs
@@ -1,15 +1,8 @@
-mod common;
 mod compositor_instance;
-mod output_receiver;
-mod packet_sender;
-mod validation;
-mod video;
+mod helpers;
 
 #[cfg(test)]
 mod tests;
 
-pub use common::*;
 pub use compositor_instance::*;
-pub use output_receiver::*;
-pub use packet_sender::*;
-pub use validation::*;
+pub use helpers::*;
